refactor: centralise all dnd libs
