+++ packages/core/lib/dnd-kit/collision/dynamic/get-midpoint-impact.ts
