+++ packages/core/lib/dnd-kit/collision/dynamic/index.ts
