+++ packages/core/lib/dnd-kit/collision/dynamic/store.ts
